Powered by OwnROM!
Also, remove progress spinner.
When booting, new apps are optimized with dex2oat. During this process a
progress dialog is displayed (optimizing app 13/74). This dialog has an
indetermninate progress spinner that uses 50% of the cpu.

This leaves only 50% for dex2oat, prolonging the time to boot, so let's remove it.

Signed-off-by: ROM-Jeremy <Disciple97414@Gmail.com>